LPS-130359 Refactor to fix shared state bug
As explained here:

    https://github.com/liferay-lima/liferay-portal/pull/1748#issuecomment-821050326

`blogs-web` has two `<liferay-item-selector:image-selector>` taglibs in
its `edit_entry.jsp`. Only the one the cover image one has a caption;
the "small image" under "Configuration" does not.

So, when we were firing events with `Liferay.fire()` from _either_
selector, it would wind up hiding or showing the caption on the cover
image. This is a bit of a smell anyway, because the image selector is
supposed to be generic, but it was secretly coupled to `blogs-web` and
was sending it a special event, and doing so in a non-specific way (ie.
an event without an id, which meant that `blogs-web` couldn't
distinguish between cover image changes and small image changes).

The fix, then, is to go back to managing `image` state internally to the
component, and only communicate with `blogs-web` for the case of the
cover image (as identified by the `paramName` attribute). The bug is
gone, although a bit of the smell remains (ie. the generic selector is
still doing something secret/special for blogs). The alternative of
making the secret "public" by adding an attribute to the taglib isn't
much of an improvement, so we're going with this approach as it is
minimal and it fixes the bug.

It also means that this isn't an amazing showcase of `Liferay.State`,
but it does at least show that you can use it to replace
`Liferay.fire()`/`Liferay.on()`, and that you can adopt TS in an
existing module even in the smallest of ways, if you want to. The state
atom still contains the image info (`fileEntryId` and `src`), even
though something smaller would have worked (eg. a boolean). It's still
pretty small, and at least encodes some information that might be useful
when debugging.

Test plan:

1. Before this commit (and on `master`). See that adding or removing an
   image to the cover image selector or the small image selector causes
   the caption to appear disappear.
2. After this commit, see that adding/removing works; the caption only
   shows/hides in response to cover image changes. Small and cover
   images can be added/removed without affecting the other.