LPS-131310 Avoid duplicate headers
So, trying to fix the problem with duplicate headers by using some
settings documented in the `eslint-plugin-notice` README:

    https://www.npmjs.com/package/eslint-plugin-notice

The root cause for most of these is that a file was copied from the
`modules/apps/` directory into `modules/dxp/`, or vice versa. As the
plugin was configured to be dumb/conservative, it just checks to see if
the header matches expectations, and if it does not, it _prepends_ the
correct header.

So, in the case of copying a file from `modules/apps/` into
`modules/dxp/`, we wind up with a file that has an enterprise license
header followed by a GPL header. When copying in the opposite direction,
we get the headers in the opposite order.

Fix then, is to set it to replace rather than prepend. We didn't do this
originally because we were a bit scared that it might be a bit
destructive, but I've run it across all of `liferay-portal` and it
didn't mangle anything in an untoward way.

I found all the JS files with dupe headers with:

    git grep --all-match -e "This library is free" --or -e "Subscription License" -- '**/*.js'

I then removed the duplicates manually. That's what you see in this
commit.

Finally, to test that the lint rule works as expected, I went back into
those files and _swapped_ the headers (ie. putting the GPL in instead of
the enterprise license under `modules/dxp/`, and putting the enterprise
license instead of the GPL under `modules/`). Then I ran `yarn format`
and saw that everything was corrected back to the desired state (ie.
`git diff` produced no output).

One final observation: the reason we're not baking this into the default
config that comes in via `@liferay/npm-scripts` is that we send that to
potentially multiple branches that may want to have their own
`notice/notice` config, and we don't want to hard-code too much
knowledge of DXP-specific paths into that package.